+++ tests/sysvars/programs/sysvars/src/lib.rs
@@ -1,0 +1,18 @@
+use anchor_lang::prelude::*;
+
+declare_id!("Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS");
+
+#[program]
+mod sysvars {
+    use super::*;
+    pub fn sysvars(_ctx: Context<Sysvars>) -> ProgramResult {
+        Ok(())
+    }
+}
+
+#[derive(Accounts)]
+pub struct Sysvars<'info> {
+    pub clock: Sysvar<'info, Clock>,
+    pub rent: Sysvar<'info, Rent>,
+    pub stake_history: Sysvar<'info, StakeHistory>,
+}